chore: Add state module for migrated accounts
Added a new module called `state` in the `sonic_account_migrater` directory. This module contains the definition of the `MigratedAccountsState` enum and the `MigratedAccount` struct. These types represent the different states of the program accounts and the details of each migrated account.

Refactor the `instruction` module

Refactored the `instruction` module in the `sonic_account_migrater` directory. Removed the `migrate_remote_accounts` and `deactivate_remote_accounts` functions and replaced them with new variants of the `ProgramInstruction` enum. The new variants, `MigrateRemoteAccounts` and `DeactivateRemoteAccounts`, now include a vector of addresses as a parameter.

Update `mod.rs` file

Added a new module called `migrated_accounts` in the `sonic_account_migrater` directory. This module is used to declare the ID for the `MigratedAccounts` program account.