refactor: Update keypair_base58 to keypair_file in Config struct
This commit updates the `Config` struct in `config.rs` to replace the `keypair_base58` field with the `keypair_file` field. The `keypair_file` field now stores the path to the keypair file instead of the base58-encoded keypair string. This change improves the portability and security of the application.

The `keypair_base58` field is removed, and the `keypair_file` field is added with the default value of "~/.config/solana/id.json".

This commit also updates the `send_status_to_baselayer` method in `remote_loader.rs` to read the keypair from the file specified in the `Config` struct.